moved models to upper level of hierarchy
