+++ src/main/java/com/pau_pau/project/data/controllers/accounts/AccountController.java
@@ -1,10 +1,10 @@
 package com.pau_pau.project.data.controllers.accounts;
 
 import com.pau_pau.project.data.controllers.ControllerConstants;
-import com.pau_pau.project.data.models.accounts.Account;
-import com.pau_pau.project.data.models.accounts.AccountDto;
-import com.pau_pau.project.data.models.accounts.Role;
 import com.pau_pau.project.data.services.accounts.AccountService;
+import com.pau_pau.project.models.accounts.Account;
+import com.pau_pau.project.models.accounts.AccountDto;
+import com.pau_pau.project.models.accounts.Role;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
