+++ src/main/java/com/pau_pau/project/data/services/accounts/AccountServiceImpl.java
@@ -1,9 +1,9 @@
 package com.pau_pau.project.data.services.accounts;
 
 import com.pau_pau.project.common.utils.PasswordEncoderUtil;
-import com.pau_pau.project.data.models.accounts.Account;
-import com.pau_pau.project.data.models.accounts.Role;
 import com.pau_pau.project.data.repository.accounts.AccountsRepository;
+import com.pau_pau.project.models.accounts.Account;
+import com.pau_pau.project.models.accounts.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
