+++ src/main/java/com/pau_pau/project/data/services/accounts/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.pau_pau.project.data.services.accounts;
 
-import com.pau_pau.project.data.models.accounts.Account;
 import com.pau_pau.project.data.repository.accounts.AccountsRepository;
+import com.pau_pau.project.models.accounts.Account;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
