+++ src/main/java/com/pau_pau/project/data/services/directors/DirectorsServiceImpl.java
@@ -1,8 +1,8 @@
 package com.pau_pau.project.data.services.directors;
 
-import com.pau_pau.project.data.models.directors.Director;
-import com.pau_pau.project.data.models.directors.DirectorDTO;
 import com.pau_pau.project.data.repository.directors.DirectorsRepository;
+import com.pau_pau.project.models.directors.Director;
+import com.pau_pau.project.models.directors.DirectorDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
