+++ src/main/java/com/pau_pau/project/data/services/films/FilmsServiceImpl.java
@@ -1,8 +1,8 @@
 package com.pau_pau.project.data.services.films;
 
-import com.pau_pau.project.data.models.films.Film;
-import com.pau_pau.project.data.models.films.FilmDTO;
 import com.pau_pau.project.data.repository.films.FilmsRepository;
+import com.pau_pau.project.models.films.Film;
+import com.pau_pau.project.models.films.FilmDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
