+++ src/main/java/com/pau_pau/project/models/accounts/Account.java
@@ -1,4 +1,4 @@
-package com.pau_pau.project.data.models.accounts;
+package com.pau_pau.project.models.accounts;
 
 import javax.persistence.*;
