+++ src/main/java/com/pau_pau/project/models/films/Film.java
@@ -1,9 +1,9 @@
-package com.pau_pau.project.data.models.films;
+package com.pau_pau.project.models.films;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import com.pau_pau.project.data.models.directors.Director;
-import com.pau_pau.project.data.models.directors.DirectorDTO;
+import com.pau_pau.project.models.directors.Director;
+import com.pau_pau.project.models.directors.DirectorDTO;
 
 import javax.persistence.*;
 import java.util.Date;
