+++ src/main/java/com/pau_pau/project/models/films/FilmDTO.java
@@ -1,7 +1,7 @@
-package com.pau_pau.project.data.models.films;
+package com.pau_pau.project.models.films;
 
-import com.pau_pau.project.data.models.directors.Director;
-import com.pau_pau.project.data.models.directors.DirectorDTO;
+import com.pau_pau.project.models.directors.Director;
+import com.pau_pau.project.models.directors.DirectorDTO;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.sql.Timestamp;
